61 backend  actualizacion limpieza de imagenes

+++ app-ia/utils/file_ops.py
@@ -40,5 +40,34 @@ def copy_db_image_to_frontend(image_path: str) -> str:
     os.makedirs(carpeta_destino, exist_ok=True)
     nombre_foto = os.path.basename(image_path)
     destino = os.path.join(carpeta_destino, nombre_foto)
-    shutil.copy(image_path, destino)
+    if not os.path.exists(destino):
+        shutil.copy(image_path, destino)
     return nombre_foto
+
+def delete_uploaded_imagen(path_uploaded: str):
+    """
+    Elimina la imagen una vez que se obtienen los embeddings.
+
+    Args:
+        path_upload: Nombre de imagen guardada en uploads.
+    """
+    os.remove(path_uploaded)
+    print("Imagen removida")
+
+def update_recientes(path_reciente, rut: str):
+    """
+    Se actualiza la carpeta recientes de la entidad correspondiente.
+
+    Args:
+        path_reciente: Nombre de la carpeta reciente de la entidad correspondiente.
+    """
+    carpeta_recientes = os.path.join('..', 'data', 'recientes', rut)
+    archivos = [f for f in os.listdir(carpeta_recientes) if os.path.isfile(os.path.join(carpeta_recientes, f))]
+    
+    if len(archivos) < 5:
+        shutil.copy(path_reciente, carpeta_recientes)
+    else:
+        #aquí falta ver cada cuanto tiempo se aceptará una nueva foto
+        archivos.sort(key=lambda x: os.path.getmtime(os.path.join(carpeta_recientes, x)))
+        os.remove(os.path.join(carpeta_recientes, archivos[0]))
+        shutil.copy(path_reciente, carpeta_recientes)
